+++ src/test/java/lotto/ApplicationTest.java
@@ -58,18 +58,20 @@ class ApplicationTest extends NsTest {
     void 일등_테스트() {
         assertRandomUniqueNumbersInRangeTest(
                 () -> {
-                    run("1000", "6,7,9,11,17,18", "45");
+                    run("2000", "6,7,9,11,17,18", "45");
                     assertThat(output()).contains(
-                            "1개를 구매했습니다.",
+                            "2개를 구매했습니다.",
                             "[6, 7, 9, 11, 17, 18]",
                             "3개 일치 (5,000원) - 0개",
                             "4개 일치 (50,000원) - 0개",
                             "5개 일치 (1,500,000원) - 0개",
                             "5개 일치, 보너스 볼 일치 (30,000,000원) - 0개",
-                            "6개 일치 (2,000,000,000원) - 1개"
+                            "6개 일치 (2,000,000,000원) - 1개",
+                            "총 수익률은 100000000.0%입니다."
                     );
                 },
-                List.of(6, 7, 9, 11, 17, 18)
+                List.of(6, 7, 9, 11, 17, 18),
+                List.of(1, 2, 3, 4, 5, 8)
         );
     }
