+++ src/test/java/lotto/ticket/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.ticket;
 
 import lotto.ticket.Lotto;
 import org.junit.jupiter.api.DisplayName;
