+++ src/test/java/lotto/ticket/WinningTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.ticket;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
